// src/app/interfaces/nufi.interface.ts
export interface RESTReportenufiData {
    ReportId:     string;
    Status:       string;
    ScoreReporte: string;
    Company:      null;
    Person:       Person;
}

export interface Person {
    Curp_Consulta:ConsultaCurp;
    RFC_Consulta:Consultarfc;
    RFC_Estatus:Estatusrfc;
    CertificadoSat_Consulta:Certificadosat;
    Siger_Consulta:Sigerconsulta;
    CedulaProfesional_Consulta:Consultacedula;
    Rug_GarantiasMobiliarias:null;
    ListasNegras_Internacionales:null;
    ContribuyentesBoletinados69B_Consulta:Listanegrasat69B;
    Ocr_Frente:                                   null;
    Ocr_Reverso:                                  null;
    ListaNominal_Consulta:                        null;
    Geolocalizacion:                              null;
    MultiBuro_ReportComplete:                     null;
    AntecedentesJudiciales_PersonaFisicaNacional:  Listanegranacional;
    SeguridadSocial_ConsultarResponse: null;
    SeguridadSocial_HistorialResponse: null;
    Infonavit_Consulta: null;
    Enriquecimiento_Datos: null;
    AnalisisYPerfilamiento_Datos: null;
    Resultados_Google: null;
    Resultados_Noticias_Google: null;
}

export interface ConsultaCurp {
    status:  string;
    message: string;
    data:    Data | null;
    code:    number;
}
// ...
export interface Consultarfc {
    code:    number;
    status:  string;
    message: string;
    data:    Datarfc | null;
}

export interface Datarfc {
    rfc: string;
}

export interface Estatusrfc {
    status:  string;
    code:    string;
    message: string;
    data:    Dataestatus | null;
}

export interface Dataestatus {
    estatus: string;
}

export interface Certificadosat {
    status:  string;
    code:    number;
    message: string;
    data:    Datacerti | null;
}

export interface Datacerti {
    rfc:          string;
    razon_social: string;
    certificados: Certificado[];
}

export interface Certificado {
    numero_serie:  string;
    estado:        string;
    tipo:          string;
    fecha_inicial: string;
    fecha_final:   string;
    certificado:   string;
}
export interface Sigerconsulta {
    status:  string;
    code:    number;
    message: string;
    count:   number;
    data:    Datum[];
}
// ...
export interface Consultacedula {
    status:  string;
    code:    number;
    mensaje: string;
    data:    Datacedula[];
}

export interface Datacedula {
    nombre:        string;
    numCedula:     string;
    profesion:     string;
    genero:        string;
    institucion:   string;
    fechaRegistro: string;
    tipo:          string;
    estado:        string;
}

export interface Listanegrasat69B {
    status:  string;
    message: string;
    data:    null;
    code:    number;
}
// Generated by https://quicktype.io

export interface Listanegranacional {
    code:    number;
    status:  string;
    message: string;
    data:    Datanegranacional;
}